# backend/src/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Medilink Healthcare API",
    "description": "Comprehensive Interactive API Documentation for Medilink Healthcare Platform",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Local Development Server"
    },
    {
      "url": "/",
      "description": "Relative server route"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Supabase JWT authentication token."
      }
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
...
    "/api/pharmacy/inventory": {
      "get": {
        "summary": "Get Inventory Details",
        "tags": ["Pharmacy"],
        "responses": {
          "200": { "description": "Success" }
        }
      },
      "patch": {
        "summary": "Update Stock Levels",
        "tags": ["Pharmacy"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["medicine_id", "quantity"],
                "properties": {
                  "medicine_id": { "type": "string", "format": "uuid", "example": "cc0c4e12-3b2d-4f1a-8c5e-9a1b2c3d4e5f" },
                  "quantity": { "type": "integer", "example": 50 }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Success" }
        }
      }
    },
    "/api/pharmacy/orders": {
      "get": {
        "summary": "List Pharmacy Orders",
        "tags": ["Pharmacy"],
        "responses": {
          "200": { "description": "Success" }
        }
      },
      "patch": {
        "summary": "Update Pharmacy Order Status",
        "tags": ["Pharmacy"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["order_id", "status"],
                "properties": {
                  "order_id": { "type": "string", "format": "uuid", "example": "cc0c4e12-3b2d-4f1a-8c5e-9a1b2c3d4e5f" },
                  "status": { "type": "string", "example": "DISPATCHED" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Success" }
        }
      }
    },
    "/api/pharmacy/queue": {
      "get": {
        "summary": "Get Pharmacy Queue",
        "tags": ["Pharmacy"],
        "responses": {
          "200": { "description": "Success" }
        }
      },
      "patch": {
        "summary": "Update Pharmacy Queue Item",
        "tags": ["Pharmacy"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["queue_id", "status"],
                "properties": {
                  "queue_id": { "type": "string", "format": "uuid", "example": "cc0c4e12-3b2d-4f1a-8c5e-9a1b2c3d4e5f" },
                  "status": { "type": "string", "example": "COMPLETED" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Success" }
        }
      }
    },
    "/api/pharmacy/vendors": {
      "get": {
        "summary": "List Pharmacy Vendors",
        "tags": ["Pharmacy"],
        "responses": {
          "200": { "description": "Success" }
        }
      },
      "post": {
        "summary": "Register Vendor",
        "tags": ["Pharmacy"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "contact_person", "phone"],
                "properties": {
                  "name": { "type": "string", "example": "MedSupply Corp" },
                  "contact_person": { "type": "string", "example": "Ayesha Sana" },
                  "phone": { "type": "string", "example": "9123456780" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Success" }
        }
      }
    },
    "/api/pharmacy/questions": {
      "post": {
        "summary": "Submit Pharmacy Query",
        "tags": ["Pharmacy"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["patient_name", "email", "query"],
                "properties": {
                  "patient_name": { "type": "string", "example": "Jhansi Sana" },
                  "email": { "type": "string", "example": "jhansi@medilink.io" },
                  "query": { "type": "string", "example": "Do you have stock of Amoxicillin 250mg?" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Success" }
        }
      }
    },
...
  }
}